requirements.txt
Django==3.0.8
psycopg2-binary==2.8.5
celery==4.4.6
djangorestframework==3.10.3
drf-yasg
